Adding more functionality to lobby
I added some more functionality to our lobby system. Players will now see a list
of all the current users in the same room. When a player leaves, this will
be reflected on that list. Players can add a name for themselves, and information
of these players are saved into an array. Added comments explaining committed